+++ compiler/lib/mlvalue.mli
@@ -18,7 +18,12 @@
  *)
 
 module Block : sig
-  val make : tag:int -> args:Javascript.element list -> Javascript.expression
+  val make :
+       ?cls:Javascript.expression
+    -> tag:int
+    -> args:Javascript.element list
+    -> unit
+    -> Javascript.expression
 
   val tag : Javascript.expression -> Javascript.expression
 
@@ -26,7 +31,12 @@ module Block : sig
 end
 
 module Array : sig
-  val make : tag:int -> args:Javascript.element list -> Javascript.expression
+  val make :
+       ?cls:Javascript.expression
+    -> tag:int
+    -> args:Javascript.element list
+    -> unit
+    -> Javascript.expression
 
   val length : Javascript.expression -> Javascript.expression